Split up transport planning into files
